vaik_text_generator/requirements.txt
Pillow>=9.4.0
numpy>=1.24.1